refs #77 Changed directory structure and grouped fitbenchmarking functions into smaller modules.
