+++ fitbenchmarking/data_import/__init__.py
@@ -1,0 +1,1 @@
+
