+++ fitbenchmarking/plotting/plotHelper.py
