+++ fitbenchmarking/tests/ResultsOutputTest.py
