+++ fitbenchmarking/tests/getProblemFilesTests.py
