+++ fitbenchmarking/utils/color_definitions.txt
