+++ fitbenchmarking/utils/test_result.py
