+++ fitbenchmarking/utils/utils_fit.py
@@ -1,0 +1,40 @@
+"""
+Utility functions for fit benchmarking a problem.
+"""
+# Copyright &copy; 2016 ISIS Rutherford Appleton Laboratory, NScD
+# Oak Ridge National Laboratory & European Spallation Source
+#
+# This file is part of Mantid.
+# Mantid is free software; you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation; either version 3 of the License, or
+# (at your option) any later version.
+#
+# Mantid is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+# File change history is stored at:
+# <https://github.com/mantidproject/fitbenchmarking>.
+# Code Documentation is available at: <http://doxygen.mantidproject.org>
+
+import numpy as np
+from logging_setup import logger
+
+
+def calculate_chi_sq(differences):
+    """
+    """
+
+    if fit_wks:
+        chi_sq = np.sum(np.square(differences))
+    else:
+        chi_sq = np.nan
+        logger.warning("No output fit workspace")
+
+    return chi_sq
+
